requirements-guncel.txt
# Django and related packages
Django>=4.2.0
django-crispy-forms>=2.0
python-dateutil>=2.8.2

# Google Cloud related packages
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.95.0
google-cloud-storage>=2.10.0
google-cloud-secret-manager>=2.16.1

# Production dependencies
gunicorn>=21.2.0
whitenoise>=6.5.0  # For static file handling

# Security
python-dotenv>=1.0.0
